+++ home-helper-core/src/main/kotlin/home/helper/core/dto/save/SaveOutput.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2022 Ryo-Kgym.
  */
 
-package home.helper.core.dto
+package home.helper.core.dto.save
 
 /**
  * 保存処理の実行結果を保持するクラスです。
